backend/requirements.txt
# Backend Dependencies for Smart Bus MARL

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Real-time Communication (WebSocket)
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Machine Learning
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging
colorlog==6.7.0

# Development
pytest==7.4.0
black==23.7.0